remove useless util and move e2e outside
