+++ src/utils/utils.test.js
@@ -1,28 +1,4 @@
-import { fixedZero, isUrl } from './utils';
-
-describe('fixedZero tests', () => {
-  it('should not pad large numbers', () => {
-    expect(fixedZero(10)).toEqual(10);
-    expect(fixedZero(11)).toEqual(11);
-    expect(fixedZero(15)).toEqual(15);
-    expect(fixedZero(20)).toEqual(20);
-    expect(fixedZero(100)).toEqual(100);
-    expect(fixedZero(1000)).toEqual(1000);
-  });
-
-  it('should pad single digit numbers and return them as string', () => {
-    expect(fixedZero(0)).toEqual('00');
-    expect(fixedZero(1)).toEqual('01');
-    expect(fixedZero(2)).toEqual('02');
-    expect(fixedZero(3)).toEqual('03');
-    expect(fixedZero(4)).toEqual('04');
-    expect(fixedZero(5)).toEqual('05');
-    expect(fixedZero(6)).toEqual('06');
-    expect(fixedZero(7)).toEqual('07');
-    expect(fixedZero(8)).toEqual('08');
-    expect(fixedZero(9)).toEqual('09');
-  });
-});
+import { isUrl } from './utils';
 
 describe('isUrl tests', () => {
   it('should return false for invalid and corner case inputs', () => {
